+++ Encoding/x265Enc.vb
@@ -53,12 +53,38 @@ Public Class x265Enc
         AfterEncoding()
     End Sub
 
+    Overloads Sub Encode(
+        passName As String,
+        commandLine As String,
+        priority As ProcessPriorityClass,
+        Optional isFirstChunk As Boolean = True)
+
+        p.Script.Synchronize(False, True, False, Nothing, isFirstChunk)
+
+        Using proc As New Proc
+            proc.Package = Package.x265
+            proc.Header = passName
+            proc.Encoding = Encoding.UTF8
+            proc.Priority = priority
+            proc.SkipString = "%] "
+
+            If commandLine.Contains("|") Then
+                proc.File = "cmd.exe"
+                proc.Arguments = "/S /C """ + commandLine + """"
+            Else
+                proc.CommandLine = commandLine
+            End If
+
+            proc.Start()
+        End Using
+    End Sub
+
     Overrides Function GetFixedBitrate() As Integer
         Return CInt(Params.Bitrate.Value)
     End Function
 
     Overrides Function CanChunkEncode() As Boolean
-        Return CInt(Params.Chunks.Value) <> 1
+        Return CInt(Params.Chunks.Value) > 1
     End Function
 
     Overrides Function GetChunkEncodeActions() As List(Of Action)
@@ -85,49 +111,28 @@ Public Class x265Enc
             If Params.Mode.Value = x265RateMode.TwoPass Then
                 ret.Add(Sub()
                             Encode("Video encoding pass 1" + name.Replace("_chunk", " chunk "),
-                                   GetArgs(1, startFrame, endFrame, name, p.Script), s.ProcessPriority)
+                                   GetArgs(1, startFrame, endFrame, name, p.Script), s.ProcessPriority, isFirst)
                             Encode("Video encoding pass 2" + name.Replace("_chunk", " chunk "),
-                                   GetArgs(2, startFrame, endFrame, name, p.Script), s.ProcessPriority)
+                                   GetArgs(2, startFrame, endFrame, name, p.Script), s.ProcessPriority, isFirst)
                         End Sub)
             ElseIf Params.Mode.Value = x265RateMode.ThreePass Then
                 ret.Add(Sub()
                             'Specific order 1 > 3 > 2 is correct!
                             Encode("Video encoding first pass" + name.Replace("_chunk", " chunk "),
-                                   GetArgs(1, startFrame, endFrame, name, p.Script), s.ProcessPriority)
+                                   GetArgs(1, startFrame, endFrame, name, p.Script), s.ProcessPriority, isFirst)
                             Encode("Video encoding Nth pass" + name.Replace("_chunk", " chunk "),
-                                   GetArgs(3, startFrame, endFrame, name, p.Script), s.ProcessPriority)
+                                   GetArgs(3, startFrame, endFrame, name, p.Script), s.ProcessPriority, isFirst)
                             Encode("Video encoding last pass" + name.Replace("_chunk", " chunk "),
-                                   GetArgs(2, startFrame, endFrame, name, p.Script), s.ProcessPriority)
+                                   GetArgs(2, startFrame, endFrame, name, p.Script), s.ProcessPriority, isFirst)
                         End Sub)
             Else
                 ret.Add(Sub() Encode("Video encoding" + name.Replace("_chunk", " chunk "),
-                    GetArgs(1, startFrame, endFrame, name, p.Script), s.ProcessPriority))
+                    GetArgs(1, startFrame, endFrame, name, p.Script), s.ProcessPriority, isFirst))
             End If
         Next
 
         Return ret
     End Function
-
-    Overloads Sub Encode(passName As String, commandLine As String, priority As ProcessPriorityClass)
-        p.Script.Synchronize()
-
-        Using proc As New Proc
-            proc.Package = Package.x265
-            proc.Header = passName
-            proc.Encoding = Encoding.UTF8
-            proc.Priority = priority
-            proc.SkipString = "%] "
-
-            If commandLine.Contains("|") Then
-                proc.File = "cmd.exe"
-                proc.Arguments = "/S /C """ + commandLine + """"
-            Else
-                proc.CommandLine = commandLine
-            End If
-
-            proc.Start()
-        End Using
-    End Sub
 
     Overrides Sub RunCompCheck()
         If Not g.VerifyRequirements OrElse Not g.IsValidSource Then
@@ -997,7 +1002,7 @@ Public Class x265Params
                     Scenecut,
                     New NumParam() With {.Switch = "--scenecut-bias", .Text = "Scenecut Bias", .Init = 5, .Config = {0, 100, 1, 1}},
                     New NumParam() With {.Switch = "--radl", .Text = "Radl"},
-                    New NumParam() With {.Switch = "--hist-threshold", .Text = "Hist Threshold", .Init = 0.01, .Config = {0, 2, 0.01, 2}},
+                    New NumParam() With {.Switch = "--hist-threshold", .Text = "Hist Threshold", .Init = 0.03, .Config = {0, 1, 0.01, 2}},
                     Ref)
                 Add("Slice Decision 2", MinKeyint, Keyint, Bpyramid, OpenGop, IntraRefresh,
                     New BoolParam() With {.Switch = "--fades", .Text = "Detection and handling of fade-in regions"},